feat: implement task service layer

+++ internal/domain/task.go
@@ -5,5 +5,5 @@ type Task struct {
 	Code        string
 	Title       string
 	Status      TaskStatus
-	NamespaceID int
+	NamespaceID int64
 }
